chore: Fix chromedriver_wrapper installation
Previously, chromedriver_wrapper.py, while having execute permission
in the repository, wasn't being installed with execute permission by
setuptools.  Consequently, it worked when the package was installed in
"develop" mode but not when installed normally.

With this change, it is installed using the setuptools entry_points
console_scripts mechanism to ensure it is executable.

This fixes #4.