amazon: don't download pre-order invoices
Invoices for pre-orders (i.e. items that haven't yet been shipped) look like
the payment card was charged on the date of the invoice, but this isn't
accurate.  Pre-orders also often change prices between when the order is made
and when it's shipped. All in all it yields better results (especially for
beancount-import) to wait to download them until the order is fulfilled and
the card is actually charged; that way, the invoice matches the actual history
of one's payment account.

This change by default disables the downloading of pre-order invoices.  Users
who prefer to download them immediately can set the
`download_preorder_invoices` option in their source configuration.