pass overwrite to atomic_write
I need this because I keep my finance-dl data in Google Drive mounted with
google-drive-fuse, and it doesn't support the operation atomic_write does
without it. (With overwrite=True, atomicwrites does a rename; otherwise it
does a link followed by an unlink. google-drive-fuse supports rename just
fine, but not hard links.)

The real solution is probably to fix atomicwrites to properly support all
filesystems that don't support links instead of just Windows, but this is less work.