ofx: support configurable destination directory per account
Account numbers aren't a great identifier for humans; if one wants to be able
to browse the data fetched by finance-dl, it's helpful to have mnemonic names
for the directories holding data for a given account. Even failing this, some
banks use account numbers that are not valid directory names, and users may
prefer a different method to clean the generated name than the default of
substituting all invalid characters with hyphens.

This change allows the user to add an `acct_dir_map` entry to the finance-dl
config for an ofx source. The map is a simple dict mapping account number (as
given in the OFX output) to directory name.

If a mapping is given, it is trusted blindly; if the user specifies a
directory for an account that makes the path invalid, it's better to fail than
silently do something different from what we were explicitly asked for.