ofx: try actual specified lower bound before binary searching to find earliest available start date
The ofx source tries to guess when data for an account begins. It has the user
specify a minimum start date as an anchor, and then uses binary search to find
the earliest date since then which generates a nonempty response from the
institution.

If the institution actually has data from the minimum start date, this results
in a rather silly search—and, in fact, the binary search as implemented never
tries the lower bound of the range at all (the interval being searched is open
at both ends), so if a user specifies the actual first day for which they have
data as min_start_date, finance-dl would always miss that first day. Given
that the user knowing when their data starts is at least a plausible case, we
should try it before resorting to cleverness; this change makes us do so.