amazon: refresh login as needed
Amazon seems to spontaneously require reauthentication when finance-dl tries
to download some of my invoices (particularly digital ones, it seems). I guess
something is probably detecting that finance-dl is not a human and getting
tetchy about it.

This change allows us to re-enter the password when prompted to in the
invoice-downloading flow. It also adds a couple extra terminating conditions
to the detector for a fully loaded invoice in the event of a cancelled order
or a permanent error.