amazon: always click Your Orders link
The amazon source has two config parameters `regular` and `digital` which
control fetching of invoices for regular and digital orders respectively. When
`regular` was false, we were not clicking the Your Orders link on the landing
page; this is required when fetching digital orders (because the Digital
Orders link appears on the Your Orders page but not the front page).

This change makes the amazon source always click the Your Orders link before
retrieving invoices, even if only retrieving invoices for digital orders.