chore: vendor some ofx utils from beancount-ingest
The ofx source makes use of two simple functions from an OFX importer that
used to be included in Beancount: one to parse the time format used in OFX,
and one to extract a named child node from an XML element. Unfortunately that
importer has been demoted to a separate repository beangulp, and to the
examples directory in that repository to boot. This change therefore copies
the (trivial) implementations of find_child and parse_ofx_time from the
beangulp examples into ofx.py.