healthequity: ignore balance-after when merging (#55)
The healthequity source merges newly-downloaded transaction data into the
previously-saved file. This failed when there were multiple transactions of
the same type in a single day, because the order in which HealthEquity reports
transactions is not stable and so the running balances were not either. The
result was that past transactions would sometimes be spontaneously duplicated
in the list upon a new finance-dl run.

This change causes the merge process to ignore the "Balance After" column.
This also means that the running balance within a day may end up incorrect,
if newly-available transactions happen to be listed before
previously-available ones in the new download. There's no really good way to
prevent this except either recalculating the balance-after column ourselves
after the merge or throwing it out entirely, neither of which is proposed in
this change.